Refactoring der Struktur des Projekts
