+++ src/main/java/mops/application/services/PublicationService.java
@@ -1,0 +1,8 @@
+package mops.application.services;
+
+import mops.domain.models.datepoll.DatePoll;
+
+public interface PublicationService {
+
+    public boolean publishDatePoll(DatePoll datePoll);
+}
