+++ src/main/java/mops/application/services/dummy/DummyUserService.java
@@ -1,6 +1,7 @@
-package mops.application.services;
+package mops.application.services.dummy;
 
 import lombok.NoArgsConstructor;
+import mops.application.services.UserService;
 import mops.domain.models.user.UserId;
 import org.springframework.stereotype.Service;
