+++ src/main/java/mops/application/services/implementation/DatePollSyndicationService.java
@@ -1,4 +1,4 @@
-package mops.application.services;
+package mops.application.services.implementation;
 
 import mops.domain.models.datepoll.DatePoll;
 import mops.domain.models.datepoll.DatePollBuilder;
