+++ src/main/java/mops/infrastructure/adapters/webflow/datepoll/dtos/ConfirmationDto.java
@@ -1,0 +1,7 @@
+package mops.infrastructure.adapters.webflow.datepoll.dtos;
+
+import java.io.Serializable;
+
+public class ConfirmationDto implements Serializable {
+    // TODO: füllen
+}
