+++ src/main/java/mops/infrastructure/config/KeycloakConfig.java
@@ -1,4 +1,4 @@
-package mops.config;
+package mops.infrastructure.config;
 
 import org.keycloak.adapters.springboot.KeycloakSpringBootConfigResolver;
 import org.springframework.context.annotation.Bean;
