+++ src/main/resources/flows/errormappings/publicationmappings.properties
